spark/requirements.txt
kafka-python
pyspark
python-dotenv
spacy